graphene_example/pyproject.toml
[tool.poetry]
name = "graphene_example"
version = "0.1.0"
description = "FastAPI Starlette Graphene integration example"
authors = ["Danil Maximov <bazinga.mail@yandex.ru>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
starlette = "^0.14.2"
fastapi = "^0.68.0"
graphene = "^3.0"
sqlmodel = "^0.0.4"
psycopg2 = "^2.9.2"
graphql-relay = "^3.1.0"
python-dotenv = "^0.19.2"
uvicorn = "^0.16.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
